+++ view/frontend/templates/hyva/recommendations.phtml
@@ -17,23 +17,25 @@
 
 declare(strict_types=1);
 
-use ImaginationMedia\AlgoliaProductRecommendations\ViewModel\ProductRecommendationsViewModel;
+use ImaginationMedia\AlgoliaProductRecommendations\ViewModel\RecommendationsViewModel;
 use Magento\Framework\Escaper;
 use Magento\Framework\View\Element\Template;
 
 /** @var Template $block */
 /** @var Escaper $escaper */
-/** @var ProductRecommendationsViewModel $relatedProducts */
-$relatedProducts = $block->getData('relatedProductsViewModel');
+/** @var RecommendationsViewModel $recommendationsViewModel */
+$recommendationsViewModel = $block->getData('recommendationsViewModel');
+$uniqueId = $block->getData('uniqueId');
 
-if (!$relatedProducts->isEnabled()) {
-    return;
+if (!$recommendationsViewModel->isEnabled()) {
+    return ;
 }
 ?>
 
 <?= $block->getChildHtml('js') ?>
 
-<div x-data="algoliaRecommendationsFor<?= $escaper->escapeHtmlAttr($relatedProducts->getProductId()) ?>()"
+<div x-data="{...algoliaRecommendations(), ...algoliaRecommendationsFor<?= $escaper->escapeHtmlAttr($uniqueId) ?>()}"
+     x-spread="eventListeners"
      x-cloak
      x-init="init"
      x-show="nbHits"
@@ -41,7 +43,7 @@ if (!$relatedProducts->isEnabled()) {
     <h4 class="text-2xl font-semibold mb-5"><?= $escaper->escapeHtml(__($block->getData('title'))) ?></h4>
 
     <template x-if="nbHits">
-        <div class="grid grid-cols-2 gap-5 md:grid-cols-3 lg:grid-cols-4 xl:grid-cols-5">
+        <div class="grid grid-cols-2 gap-5 md:grid-cols-3 lg:grid-cols-5">
             <template x-for="hit in hits">
                 <?= $block->getChildHtml('hit') ?>
             </template>
